Extract MCP Codex Apps support
Move Codex Apps tool cache, cache key, connector filtering, and callable-name normalization into codex-mcp/src/apps.rs.

Update client, manager, exports, and tests to import those helpers from the new module.
