+++ codex-rs/codex-mcp/src/manager.rs
@@ -5,6 +5,9 @@ use std::time::Duration;
 use std::time::Instant;
 
 use crate::McpAuthStatusEntry;
+use crate::apps::CodexAppsToolsCacheContext;
+use crate::apps::CodexAppsToolsCacheKey;
+use crate::apps::write_cached_codex_apps_tools_if_needed;
 use crate::client::AsyncManagedClient;
 use crate::client::MCP_TOOLS_FETCH_UNCACHED_DURATION_METRIC;
 use crate::client::MCP_TOOLS_LIST_DURATION_METRIC;
@@ -14,8 +17,6 @@ use crate::client::list_tools_for_client_uncached;
 use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
-use crate::mcp_connection::CodexAppsToolsCacheContext;
-use crate::mcp_connection::CodexAppsToolsCacheKey;
 use crate::mcp_connection::McpRuntimeEnvironment;
 use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
@@ -26,7 +27,6 @@ use crate::mcp_connection::qualify_tools;
 use crate::mcp_connection::startup_outcome_error_message;
 use crate::mcp_connection::tool_with_model_visible_input_schema;
 use crate::mcp_connection::transport_origin;
-use crate::mcp_connection::write_cached_codex_apps_tools_if_needed;
 use anyhow::Context;
 use anyhow::Result;
 use anyhow::anyhow;
